Refactored bridge_data
