+++ horde-scribe-bridge.py
@@ -6,7 +6,7 @@ from worker.utils.set_envs import set_worker_env_vars_from_config
 
 set_worker_env_vars_from_config()  # Get `cache_home` from `bridgeconfig.yaml` into the environment variable
 
-from worker.bridge_data.scribe import KoboldAIBridgeData  # noqa: E402
+from worker.bridge_data import BridgeData  # noqa: E402
 from worker.logger import logger, quiesce_logger, set_logger_verbosity  # noqa: E402
 from worker.workers.scribe import ScribeWorker  # noqa: E402
 
@@ -14,7 +14,7 @@ from worker.workers.scribe import ScribeWorker  # noqa: E402
 def main():
     set_logger_verbosity(args.verbosity)
     quiesce_logger(args.quiet)
-    bridge_data = KoboldAIBridgeData()
+    bridge_data = BridgeData()
     bridge_data.reload_data()
     try:
         worker = ScribeWorker(bridge_data)
