+++ worker/consts.py
@@ -1,3 +1,3 @@
-BRIDGE_VERSION = 24
-RELEASE = f"{BRIDGE_VERSION}.1.0"
+BRIDGE_VERSION = 25
+RELEASE = f"{BRIDGE_VERSION}.1.1"
 BRIDGE_CONFIG_FILE = "bridgeData.yaml"
